Cloudwatch - Do not allow NaN values
